tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/async_011.htm

<!DOCTYPE html>
<html>
    <head>
        <title>An empty parser-inserted script element should return async=true</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta description="An empty parser-inserted script element should return async=true." />
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
        <link rel="help" href="https://html.spec.whatwg.org/multipage/#prepare-a-script"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <script></script>
        <script type="text/javascript">
            test(function() {   assert_true(document.getElementsByTagName("script")[2].async)}, "An empty parser-inserted script element should return async=true");
        </script>
    </body>
</html>
